+++ CHANGES.rst
@@ -10,6 +10,8 @@ Changelog
 
 **Changed**
 
+- #1076 Instrument QC Viewlet Availability
+- #1071 Reinvented Listing Tables
 - #1066 Set default page size for listings to 50
 - #1063 Permission for ar_add changed to "AddAnalysisRequest"
 - #1064 Python 2.x is not supported by WeasyPrint v43. Pinned version: 0.42.3